Refactor render_crop logic into separate file, improve code quality and comments
