+++ app/controllers/alchemy/admin/essence_pictures_controller.rb
@@ -108,29 +108,20 @@ module Alchemy
         end
       end
 
-      # Combines separated gravity params (size, x, y) into single gravity hash
+      # Remove any blanks from render_gravity param hash,
+      # if all blank, set render_gravity=nil
       #
-      def combine_gravities(essence_picture_params)
-        gravity = @essence_picture.render_gravity.presence || {}
-        gravity = gravity.symbolize_keys.merge({
-          size: essence_picture_params.delete(:render_gravity_size),
-          x: essence_picture_params.delete(:render_gravity_x),
-          y: essence_picture_params.delete(:render_gravity_y),
-        }.compact)
+      def fix_render_gravity_params(essence_picture_params)
+        essence_picture_params[:render_gravity] = essence_picture_params[:render_gravity]
+          &.delete_if { |_k, v| v.blank? }.presence
 
-        # Remove any blanks after setting values, render_gravity=nil if all empty
-        gravity = gravity.delete_if { |_k, v| v.blank? }.presence
-
-        essence_picture_params[:render_gravity] = gravity
         essence_picture_params
       end
 
       def essence_picture_params
-        essence_picture_params = params.require(:essence_picture).permit(:alt_tag, :caption, :css_class, :render_size, :title, :crop_from, :crop_size, :render_gravity, :render_gravity_size, :render_gravity_x, :render_gravity_y)
+        essence_picture_params = params.require(:essence_picture).permit(:alt_tag, :caption, :css_class, :render_size, :title, :crop_from, :crop_size, render_gravity: [:size, :x, :y])
 
-        if @essence_picture.present? && [:render_gravity_size, :render_gravity_x, :render_gravity_y].any? { |k| essence_picture_params.key?(k) }
-          essence_picture_params = combine_gravities(essence_picture_params)
-        end
+        essence_picture_params = fix_render_gravity_params(essence_picture_params)
 
         essence_picture_params
       end
