+++ app/views/alchemy/admin/essence_pictures/edit.html.erb
@@ -28,13 +28,16 @@
   <%- end -%>
   <%- if @content.settings[:gravity].present? -%>
     <%- gravity = @essence_picture.gravity -%>
-    <% [:size, :x, :y].each do |type| %>
-      <%= f.input "render_gravity_#{type}".to_sym,
-        label: Alchemy.t("gravity_#{type}"),
-        collection: @essence_picture.available_gravities[type].map{|g| [Alchemy.t(g).capitalize, g]},
-        selected: @essence_picture.render_gravity ? @essence_picture.render_gravity[type.to_s] : '',
-        include_blank: Alchemy.t("Layout default"),
-        input_html: {class: 'alchemy_selectbox'} %>
+    <%= f.simple_fields_for :render_gravity do |ff| %>
+      <% @essence_picture.available_gravities.keys.each do |type| %>
+        <%= ff.input type,
+          label: Alchemy.t("gravity_#{type}"),
+          collection: @essence_picture.available_gravities[type].map{|g| [Alchemy.t(g).capitalize, g]},
+          selected: @essence_picture.render_gravity ? @essence_picture.render_gravity[type.to_s] : '',
+          include_blank: Alchemy.t("Layout default"),
+          input_html: {class: 'alchemy_selectbox'}
+        %>
+      <% end %>
     <% end %>
   <%- end -%>
   <%= f.submit Alchemy.t(:save) %>
