+++ spec/controllers/alchemy/admin/essence_pictures_controller_spec.rb
@@ -202,7 +202,7 @@ module Alchemy
     end
 
     describe "#update" do
-      before(:each) do |test|
+      before do |test|
         expect(EssencePicture).to receive(:find).and_return(essence) unless test.metadata[:stub_essence_picture]
         expect(Content).to receive(:find).and_return(content)
       end
@@ -235,13 +235,15 @@ module Alchemy
       end
 
       context "with render_gravity params" do
-        it "combines separated render_gravitiy params into serialized json hash" do
+        it "saves a compact serialized json hash to db" do
           put :update, params: {
             id: 1,
             essence_picture: {
-              render_gravity_size: "shrink",
-              render_gravity_x: "left",
-              render_gravity_y: "top",
+              render_gravity: {
+                size: "shrink",
+                x: "left",
+                y: "",
+              },
             },
           }, xhr: true
 
@@ -249,30 +251,6 @@ module Alchemy
           expect(essence.render_gravity).to eq({
             "size" => "shrink",
             "x" => "left",
-            "y" => "top",
-          })
-        end
-
-        it "can set/unset individual gravity values", :stub_essence_picture do
-          essence = EssencePicture.new(id: 1, content: content, picture: picture, render_gravity: {size: "shrink", x: "left"})
-          allow(EssencePicture).to receive(:find).and_return(essence)
-
-          # load_essence_picture did not find our essence in before_action so i set @essence_picture manually
-          expect(controller).to receive(:load_essence_picture).and_return(nil)
-          controller.instance_variable_set(:@essence_picture, essence)
-
-          put :update, params: {
-            id: 1,
-            essence_picture: {
-              render_gravity_y: "top",
-              render_gravity_x: "",
-            },
-          }, xhr: true
-
-          expect(response.status).to eq(200)
-          expect(essence.render_gravity).to eq({
-            "size" => "shrink",
-            "y" => "top",
           })
         end
 
@@ -280,7 +258,9 @@ module Alchemy
           put :update, params: {
             id: 1,
             essence_picture: {
-              render_gravity_x: "unknown",
+              render_gravity: {
+                x: "unknown",
+              },
             },
           }, xhr: true
 
@@ -288,7 +268,7 @@ module Alchemy
           expect(essence.errors.details[:render_gravity]).to eq([{error: :invalid}])
         end
 
-        it "Sets render_gravity = nil if all render_gravities are blank", :stub_essence_picture do
+        it "sets render_gravity = nil if all render_gravities are blank", :stub_essence_picture do
           essence = EssencePicture.new(id: 1, content: content, picture: picture, render_gravity: {size: "shrink"})
           allow(EssencePicture).to receive(:find).and_return(essence)
 
@@ -299,15 +279,35 @@ module Alchemy
           put :update, params: {
             id: 1,
             essence_picture: {
-              render_gravity_size: "",
-              render_gravity_x: "",
-              render_gravity_y: "",
+              render_gravity: {
+                size: "",
+                x: "",
+                y: "",
+              },
             },
           }, xhr: true
 
           expect(response.status).to eq(200)
           expect(essence.render_gravity).to equal(nil)
         end
+      end
+    end
+
+    describe "#fix_render_gravity_params" do
+      it "removes any blanks from render_gravity" do
+        expect(
+          subject.send(:fix_render_gravity_params, {
+            render_gravity: {size: "grow", x: "", y: ""},
+          })
+        ).to eq({render_gravity: {size: "grow"}})
+      end
+
+      it "sets render_gravity=nil if all gravities blank" do
+        expect(
+          subject.send(:fix_render_gravity_params, {
+            render_gravity: {size: "", x: "", y: ""},
+          })
+        ).to eq({render_gravity: nil})
       end
     end
