+++ spec/rails_helper.rb
@@ -27,6 +27,7 @@ require_relative "factories"
 require_relative "support/calculation_examples.rb"
 require_relative "support/hint_examples.rb"
 require_relative "support/transformation_examples.rb"
+require_relative "support/render_crop_examples.rb"
 require_relative "support/capybara_helpers.rb"
 require_relative "support/custom_news_elements_finder"
