+++ spec/views/alchemy/admin/essence_pictures/edit_spec.rb
@@ -52,9 +52,9 @@ describe "alchemy/admin/essence_pictures/edit.html.erb" do
     expect(rendered).to_not have_selector(".input.essence_picture_render_size")
   end
 
-  it "displays gravity selection if present" do
+  it "displays gravity selection if gravity setting present" do
     allow(content).to receive(:settings).and_return({
-      gravity: true,
+      gravity: true, # Could also be a hash that overrides default_gravity
     })
 
     render
